backend/requirements.txt
action-msgs==2.0.3
actionlib-msgs==5.3.6
actuator-msgs==0.0.1
ament-cmake-test==2.5.4
ament-copyright==0.17.3
ament-cppcheck==0.17.3
ament-cpplint==0.17.3
ament-flake8==0.17.3
ament-index-python==1.8.1
ament-lint==0.17.3
ament-lint-cmake==0.17.3
ament-package==0.16.4
ament-pep257==0.17.3
ament-uncrustify==0.17.3
ament-xmllint==0.17.3
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==3.2.0
builtin-interfaces==2.0.3
cachetools==6.2.5
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
composition-interfaces==2.0.3
cryptography==46.0.3
cv-bridge==4.1.0
deprecation==2.1.0
diagnostic-msgs==5.3.6
distro==1.9.0
dnspython==2.8.0
domain-coordinator==0.12.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.128.0
fsspec==2026.1.0
geometry-msgs==5.3.6
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.48.0rc0
google-auth-httplib2==0.3.0
google-genai==1.60.0
google-generativeai==0.3.2
googleapis-common-protos==1.72.0
gps-msgs==2.1.1
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
interactive-markers==2.5.5
laser-geometry==2.7.2
launch==3.4.9
launch-ros==0.26.10
launch-testing==3.4.9
launch-testing-ros==0.26.10
launch-xml==3.4.9
launch-yaml==3.4.9
lifecycle-msgs==2.0.3
map-msgs==2.4.1
markdown-it-py==4.0.0
mdurl==0.1.2
message-filters==4.11.9
mmh3==5.2.0
motor==3.7.1
multidict==6.7.0
nav-msgs==5.3.6
osrf-pycommon==2.1.7
packaging==26.0
passlib==1.7.4
postgrest==2.27.2
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
PyYAML==6.0.2
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.10.1
pymongo==4.16.0
pyparsing==3.3.2
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
python-qt-binding==2.2.2
qt-gui==2.7.5
qt-gui-cpp==2.7.5
qt-gui-py-common==2.7.5
rcl-interfaces==2.0.3
rclpy==7.1.6
rcutils==6.7.4
realtime==2.27.2
requests==2.32.5
resource-retriever==3.4.4
rich==14.3.1
rmw-dds-common==3.1.0
ros-gz-bridge==1.0.16
ros-gz-interfaces==1.0.16
ros-gz-sim==1.0.16
ros2action==0.32.6
ros2bag==0.26.9
ros2bag-mcap-cli==0.26.9
ros2bag-sqlite3-cli==0.26.9
ros2cli==0.32.6
ros2component==0.32.6
ros2doctor==0.32.6
ros2interface==0.32.6
ros2launch==0.26.10
ros2lifecycle==0.32.6
ros2multicast==0.32.6
ros2node==0.32.6
ros2param==0.32.6
ros2pkg==0.32.6
ros2run==0.32.6
ros2service==0.32.6
ros2topic==0.32.6
rosbag2-interfaces==0.26.9
rosbag2-py==0.26.9
rosgraph-msgs==2.0.3
rosidl-adapter==4.6.6
rosidl-cli==4.6.6
rosidl-cmake==4.6.6
rosidl-generator-c==4.6.6
rosidl-generator-cpp==4.6.6
rosidl-generator-py==0.22.2
rosidl-generator-type-description==4.6.6
rosidl-parser==4.6.6
rosidl-pycommon==4.6.6
rosidl-runtime-py==0.13.1
rosidl-typesupport-c==3.2.2
rosidl-typesupport-cpp==3.2.2
rosidl-typesupport-fastrtps-c==3.6.2
rosidl-typesupport-fastrtps-cpp==3.6.2
rosidl-typesupport-introspection-c==4.6.6
rosidl-typesupport-introspection-cpp==4.6.6
rpyutils==0.4.2
rqt-gui==1.6.2
rqt-gui-py==1.6.2
rqt-plot==1.4.4
rqt-py-common==1.6.2
rqt-topic==1.7.4
rsa==4.9.1
sensor-msgs==5.3.6
sensor-msgs-py==5.3.6
service-msgs==2.0.3
shape-msgs==5.3.6
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sros2==0.13.4
starlette==0.50.0
statistics-msgs==2.0.3
std-msgs==5.3.6
std-srvs==5.3.6
stereo-msgs==5.3.6
storage3==2.27.2
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.27.2
supabase-auth==2.27.2
supabase-functions==2.27.2
tenacity==9.1.2
tf2-geometry-msgs==0.36.16
tf2-kdl==0.36.16
tf2-msgs==0.36.16
tf2-py==0.36.16
tf2-ros-py==0.36.16
tf2-sensor-msgs==0.36.16
tf2-tools==0.36.16
theora-image-transport==4.0.6
tqdm==4.67.1
trajectory-msgs==5.3.6
type-description-interfaces==2.0.3
typing-inspection==0.4.2
typing_extensions==4.15.0
unique-identifier-msgs==2.5.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
vision-msgs==4.1.1
visualization-msgs==5.3.6
websockets==15.0.1
xacro==2.1.1
yarl==1.22.0
